pyproject.toml
[tool.poetry]
name = "open-deep-research-langgraph"
version = "0.1.0"
description = ""
authors = ["kamran945 <kamran945@gmail.com>"]
readme = "README.md"
packages = [
    { include = "src" },
    { include = "src/report_writer" }
]

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
langchain = "^0.3.7"
langchain-community = "^0.3.5"
langchain-huggingface = "^0.1.2"
langchain-groq = "^0.2.1"
tavily-python = "^0.5.0"
langgraph = "^0.2.45"
langchain-core = "^0.3.15"
python-dotenv = "^1.0.1"
langgraph-cli = {extras = ["inmem"], version = "^0.1.61"}
langchain-ollama = "^0.2.3"
gradio = "^5.17.0"
jupyterlab = "^4.3.5"
ipykernel = "^6.29.5"
duckduckgo-search = "^7.4.4"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
